AluraCreatingAPI/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docx==0.2.4
durationpy==0.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.1.0
faiss-cpu==1.8.0
Faker==19.6.1
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huepy==1.2.1
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.4
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
instabot==0.117.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==31.0.0
langchain==0.3.7
langchain-community==0.0.38
langchain-core==0.3.19
langchain-experimental==0.0.58
langchain-huggingface==0.0.3
langchain-openai==0.0.8
langchain-text-splitters==0.3.2
langchainhub==0.1.15
langsmith==0.1.125
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.0
mock==5.1.0
monotonic==1.6
moviepy==2.1.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.2
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.54.4
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
oracledb==2.4.1
orjson==3.10.7
outcome==1.2.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pd==0.0.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.6.6
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMySQL==1.1.1
pypdf==5.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.3
rich==13.9.4
robotframework==6.1.1
robotframework-faker==5.0.0
robotframework-pythonlibcore==4.2.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary @ git+https://github.com/robotframework/SeleniumLibrary.git@2f3037eeebd21de1e24db86a5d9cc0a22a04639f
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.4.3
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.5
streamlit==1.39.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
trio==0.22.2
trio-websocket==0.10.4
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.1.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.2
